+++ server/api/posts/[id].delete.ts
@@ -1,6 +1,7 @@
 import { eq } from 'drizzle-orm'
 import { db } from '../../utils/db'
 import { posts } from '../../db/schema'
+import { requireAdmin } from '../../utils/auth'
 
 export default defineEventHandler(async (event) => {
   const id = Number(getRouterParam(event, 'id'))
@@ -8,6 +9,8 @@ export default defineEventHandler(async (event) => {
   if (!id || Number.isNaN(id)) {
     throw createError({ statusCode: 400, statusMessage: 'Invalid post ID' })
   }
+
+  await requireAdmin(event)
 
   await db.delete(posts).where(eq(posts.id, id))
